Form as a Multi-page wizard
Using multiple layers of partial views, the form uses the "ui.router" package to allow multiple steps in a wizard.
app.js defines the different form states, pointing to the partials.
The parent state at /form is the only one defining a unified controller which is shared between the steps
The PartialsController needs updating whenever a new form step is added